implemented ratingbot in yasp main
