+++ public/css/yasp.css
@@ -107,6 +107,6 @@ td>img.user {
     font-weight: bolder;
 }
 
-.glyphicon{
+.big{
     font-size: 80px;
 }
